… Cython/Compiler/ParseTreeTransforms.py
… Cython/Utility/ModuleSetupCode.c
… docs/src/userguide/source_files_and_compilation.rst
… tests/compile/threadsafe_access.pyx