Closes https://github.com/cython/cython/issues/6221 (and obsoletes https://github.com/cython/cython/issues/6215)

In principle this should cover a large chunk of thread safety issues in Cython. It applies to anywhere where we hold the GIL. There's three levels:
* off
* refcounted - only applies to reference counted types (PyObject/typed memoryview)
* full - applies to all types, so read/write should at least appear "atomic".

I've made the default "refcounted", but that could be up for debate.

It has some cost in the non-free-threaded builds because:
* it unchains cdef class attribute lookups so that they go via temps.
* It makes some error handling more complicated.

Right now the main limitation is that we don't yet have access to critical sections - see https://github.com/python/cpython/issues/119344. I've put placeholders there for now, but they really don't do anything. This means that `cdef classes` and closure scopes aren't guarded properly.