@da-woods 2024-06-01
Obviously not quite working yet...

I did some crude benchmarking, just running ExprNodes.py through Cython:

Python 3.11
-----------------

This branch:  6.706s 6.764s 6.780s

Current master: 6.767s 6.721s 6.642s

pure Python: 10.823s 10.686s 10.999s

This branch seems to make a pretty negligible difference to Python 3.11.

Python 3.11 freethreading (but Python built without optimizations)
--------------------------------------

This branch: 7.575s 7.482s 7.613s

This branch and -Xgil: 7.428s 7.461s 7.550s

Current master and -Xgil=0: 6.670s 7.106s 6.836s

pure Python and -Xgil=0: 14.735s 14.196s 14.605s

-Xgil=0 makes no measurable difference here (Cython basically single-threaded so unsurprising). The extra locking does make things ~10% slower. And it may be a little worse than that since all the cdef classes aren't really being locked yet.