@da-woods 2024-06-02
I think it should be possible to eliminate a lot of the extra temps. To take full advantage though we really need critical sections (and we probably need them for in the Python 3.13, otherwise we stuck being unable to use them properly until we drop support for Python 3.13).