@da-woods 2024-06-02
Converting to draft for now...

The basic idea works quite well I think, but I'm still making changes so I don't think it'd benefit from a detailed review.