@da-woods 2024-06-09
My current thoughts here:

1. I'm waiting for a decision on "is this the right thing to do" before tidying up the remaining bugs. The remaining bugs are fairly small though.
2. In principle we could make reads from cdef classes (and closures) mostly lock free using the same method as in https://peps.python.org/pep-0703/#optimistically-avoiding-locking. That should probably be guarded by a "turn it off" flag because it's very much abusing internal details.