@scoder 2024-06-14
> It has some cost in the non-free-threaded builds because:
>
>     * it unchains cdef class attribute lookups so that they go via temps.
>     * It makes some error handling more complicated.

Is that cost verified? It seems quite possible that C compilers could undo that effect when optimising this more complex C code. Removing useless variables and collapsing or splitting code paths is something they do all the time.