@da-woods 2024-06-15
> > It has some cost in the non-free-threaded builds because:
> > ```
> > * it unchains cdef class attribute lookups so that they go via temps.
> > * It makes some error handling more complicated.
> > ```
> 
> Is that cost verified? It seems quite possible that C compilers could undo that effect when optimising this more complex C code. Removing useless variables and collapsing or splitting code paths is something they do all the time.

The temps are usually reference-counted so probably not too easy to optimize. But I couldn't measure a speed difference in Cython in non-freethreaded Python with and without this PR, so I don't think it's huge.